fix: undo halloween theme
