+++ src/components/ComponentDemo/ComponentDemo.module.scss
@@ -368,7 +368,3 @@ fieldset.form-group:last-of-type {
     height: 2.5rem;
   }
 }
-
-:global(.halloween-component-demo) ~ form .component-knob-title {
-  background-color: #6f6f6f;
-}
