+++ src/components/SVGLibraries/shared/SvgLibrary.module.scss
@@ -51,9 +51,6 @@
   top: -1rem;
   background: #f3f3f3;
 }
-:global(.halloween) .filter-row::before {
-  background: #161616;
-}
 
 .filter-row :global(.bx--search-input) {
   border-bottom: 1px solid $ui-03;
@@ -250,10 +247,6 @@ $svg-card-border: 1px solid $ui-03;
 .flex-container svg {
   fill: $icon-01;
   animation: fadeIn $duration--fast-01 motion(standard, productive);
-}
-
-:global(.halloween) .flex-container svg {
-  stroke: #161616 !important;
 }
 
 @keyframes fadeIn {
