--- src/gatsby-theme-carbon/styles/_global.scss
@@ -1,109 +1,0 @@
-.halloween {
-  //nav active styles
-  a.bx--side-nav__link--current:before,
-  a.bx--side-nav__link[aria-current='page']:before,
-  .bx--side-nav.bx--side-nav--website--light
-    .bx--side-nav__item.bx--side-nav__item--active
-    .bx--side-nav__submenu[aria-expanded='false']::before,
-  .bx--side-nav__item--active
-    .bx--side-nav__submenu[aria-expanded='false']::before {
-    background-color: #eb6200;
-  }
-
-  //page tabs
-  main.container nav li a[aria-current='page'] {
-    border-top: 3px solid #eb6200;
-  }
-
-  //focus styles
-  .bx--side-nav__submenu:focus,
-  .bx--side-nav a.bx--header__menu-item:focus,
-  a.bx--side-nav__link:focus,
-  footer a:focus,
-  .bx--tile--clickable:focus,
-  .bx--tile--expandable:focus,
-  .bx--tile:focus {
-    outline: 2px solid #eb6200;
-  }
-
-  a:focus,
-  .container--homepage a:focus,
-  .container--dark a:focus {
-    outline: 2px solid #eb6200;
-    border: none;
-  }
-
-  // tile overrides
-  .bx--resource-card {
-    border-color: black !important;
-  }
-  .bx--resource-card .bx--tile {
-    background-color: #262626;
-  }
-
-  .bx--resource-card .bx--tile:hover {
-    background-color: #393939 !important;
-  }
-
-  .bx--resource-card .bx--tile * {
-    fill: white;
-    color: white;
-  }
-
-  // codesnippet overrides
-  code,
-  pre.prism-code {
-    background-color: #262626 !important;
-  }
-
-  pre.prism-code ~ div {
-    background: #262626 !important;
-  }
-
-  code {
-    background-color: #6f6f6f !important;
-    color: white;
-  }
-
-  //labels
-  .bx--label,
-  .component-index-filter__label {
-    color: white;
-  }
-
-  //component demo codesnippet
-  .halloween-component-demo div,
-  .halloween-component-demo pre,
-  .halloween-component-demo textarea {
-    background-color: #262626 !important;
-  }
-
-  //color token table
-  .color-token-table__theme-switcher::before {
-    background: #161616;
-  }
-
-  .component-docs table tbody tr td,
-  .page-table tbody tr td {
-    color: #161616;
-  }
-
-  .halloween-callout {
-    background-color: #eb6200 !important;
-  }
-
-  .footer-logo {
-    fill: #eb6200;
-  }
-}
-
-.halloween-callout {
-  position: relative;
-  z-index: 2;
-}
-
-// Overrides for nav and text and stuff
-// https://github.com/carbon-design-system/carbon-website-archive/pull/277/files#diff-32a7ff4c1073a8e5bb8bdff4387e52d9f558676d15576269a8b1fa1b5572fd87
-
-// Components
-// https://github.com/carbon-design-system/carbon-website-archive/pull/277/files#diff-6b9601adfbd177d8d835681e286da56edb2e77e4d19657959fd1312576cb6c98
