+++ src/styles/Layout.module.scss
@@ -79,23 +79,3 @@
   color: $text-04;
   @include carbon--type-style('body-short-01');
 }
-
-.halloween-button {
-  padding-right: 12px;
-}
-
-.halloween-button:hover,
-.halloween-button:focus {
-  background-color: #eb6200;
-  color: white;
-}
-
-.halloween-button svg {
-  fill: #eb6200;
-  height: 1.5rem;
-}
-
-.halloween-button:hover svg,
-.halloween-button:focus svg {
-  fill: white;
-}
