+++ src/styles/_homepage.scss
@@ -9,6 +9,7 @@
   opacity: 0.35;
 }
 
+//make sure content is above dots
 .container--homepage div[class*='Main-module--main'],
 .container--homepage div[class*='Callout-module--grid'],
 .container--homepage div[class*='NextPrevious-module--wrapper'] {
